# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Ma premi\u00e8re API",
        "description": "Une API permettant l'acc\u00e8s aux donn\u00e9es de la base via une approche RESTful",
        "termsOfService": "",
        "contact": {
            "name": "Guiks",
            "email": "guillaume.delapre@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "myapi.com",
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/membre": {
            "get": {
                "tags": [
                    "membre"
                ],
                "summary": "Get a member list",
                "description": "Use this method to return a listing of member.",
                "operationId": "listMember",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Membre"
                            }
                        }
                    },
                    "204": {
                        "description": "The request didn't return any content."
                    }
                }
            },
            "post": {
                "tags": [
                    "membre"
                ],
                "summary": "Create a member",
                "description": "Use this method to create a new member.",
                "operationId": "createMember",
                "consumes": [
                    "multipart/form-data",
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id_personne",
                        "in": "formData",
                        "description": "Person ID of the member",
                        "required": true,
                        "type": "number",
                        "maximum": "11"
                    },
                    {
                        "name": "id_abonnement",
                        "in": "formData",
                        "description": "Subscribe ID of the member",
                        "required": true,
                        "type": "number",
                        "maximum": "11"
                    },
                    {
                        "name": "date_inscription",
                        "in": "formData",
                        "description": "Inscription date of the member",
                        "required": true,
                        "type": "string",
                        "format": "date"
                    },
                    {
                        "name": "debut_abonnement",
                        "in": "formData",
                        "description": "Initial date of the member subscription",
                        "required": true,
                        "type": "string",
                        "format": "date"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Member created",
                        "schema": {
                            "$ref": "#/definitions/Membre"
                        }
                    },
                    "422": {
                        "description": "Required fields missing or incorrect"
                    }
                }
            }
        },
        "/membre/{id_membre}": {
            "get": {
                "tags": [
                    "membre"
                ],
                "summary": "Display a single member",
                "description": "Use this method to return a single member attributes based on its id.",
                "operationId": "showMember",
                "parameters": [
                    {
                        "name": "id_membre",
                        "in": "path",
                        "description": "id of member to fetch",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Membre"
                        }
                    },
                    "404": {
                        "description": "Member not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "membre"
                ],
                "summary": "Update a member",
                "description": "Use this method to update the attributes of a member based on its id.",
                "operationId": "updateMember",
                "consumes": [
                    "multipart/form-data",
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id_membre",
                        "in": "path",
                        "description": "ID person",
                        "required": true,
                        "type": "integer",
                        "maximum": "11"
                    },
                    {
                        "name": "id_personne",
                        "in": "formData",
                        "description": "Person ID of the member",
                        "required": false,
                        "type": "number",
                        "maximum": "11"
                    },
                    {
                        "name": "id_abonnement",
                        "in": "formData",
                        "description": "Subscribe ID of the member",
                        "required": false,
                        "type": "number",
                        "maximum": "11"
                    },
                    {
                        "name": "date_inscription",
                        "in": "formData",
                        "description": "Inscription date of the member",
                        "required": false,
                        "type": "string",
                        "format": "date"
                    },
                    {
                        "name": "debut_abonnement",
                        "in": "formData",
                        "description": "Initial date of the member subscription",
                        "required": false,
                        "type": "string",
                        "format": "date"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Member updated",
                        "schema": {
                            "$ref": "#/definitions/Membre"
                        }
                    },
                    "422": {
                        "description": "Required fields missing or incorrect"
                    }
                }
            },
            "delete": {
                "tags": [
                    "membre"
                ],
                "summary": "Delete a Member",
                "description": "Use this method to delete a member based on its id.",
                "operationId": "deleteMember",
                "parameters": [
                    {
                        "name": "id_membre",
                        "in": "path",
                        "description": "Member ID to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Member deleted"
                    },
                    "404": {
                        "description": "Invalid member value"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Membre": {
            "required": [
                "id_personne",
                "id_abonnement",
                "date_inscription",
                "debut_abonnement"
            ],
            "properties": {
                "id_membre": {
                    "type": "number",
                    "format": "int64",
                    "default": 1
                },
                "id_personne": {
                    "type": "number",
                    "format": "int64",
                    "default": 1
                },
                "id_abonnement": {
                    "type": "number",
                    "format": "int64",
                    "default": 1
                },
                "date_inscription": {
                    "type": "date",
                    "format": "date",
                    "default": "2015-08-17"
                },
                "debut_abonnement": {
                    "type": "date",
                    "format": "date",
                    "default": "2015-09-01"
                }
            },
            "xml": {
                "name": "Membre"
            }
        },
...
    }
}